backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
alembic==1.12.1
psycopg2-binary==2.9.10
geoalchemy2==0.17.1
pydantic==2.11.7
pydantic-settings==2.1.0
python-dotenv==1.1.1
pytest==8.4.1
pytest-asyncio==0.21.1
httpx==0.25.2
python-multipart==0.0.6
curl_cffi==0.5.9 